(TIN-5) Refactoring
